OPEN ENDED - FARMER CHATBOT/backend/requirements.txt
Flask
openai
python-dotenv
SpeechRecognition 
pyaudio